Final fixes after testing

+++ src/PVOut.h
@@ -19,15 +19,15 @@ public:
     
     byte             size() const;
 
-    unsigned long    combineBits(unsigned long bits);
+    uint32_t         combineBits(uint32_t bits);
  
 protected:
-    unsigned long    computeBits(unsigned long);
+    uint32_t         computeBits(uint32_t);
 
     uint32_t         m_outputBits;
     byte             m_size;
     byte             m_bitPos;  
-    unsigned long    m_mask;
+    uint32_t         m_mask;
 };
 
 class GPIOOutputBank : public OutputBank
